# src/main/resources/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Verdan Skoleadministrasjon API",
    "description": "REST API for Verdan skoleadministrasjonssystem. Støtter multi-institusjon med rollebasert tilgangskontroll (SUPER_ADMIN, INSTITUTION_ADMIN, TEACHER, STUDENT).\n\nAlle endepunkter (unntatt `/api/login`, `/api/auth/refresh` og `/api/health`) krever JWT Bearer-token i `Authorization`-headeren.\n\nAccess-tokens utløper etter 15 minutter. Bruk `/api/auth/refresh` for å fornye.",
    "version": "2.0.0",
    "contact": {
      "name": "Verdan Dev Team (Gruppe 2)"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Lokal utvikling"
    },
    {
      "url": "http://localhost:8081",
      "description": "Docker"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "admin" },
          "password": { "type": "string", "example": "admin123" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login and get JWT token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT token and user info" },
          "400": { "description": "Missing username or password" },
          "401": { "description": "Invalid credentials" },
          "403": { "description": "Institution deactivated (staff only)" },
          "429": { "description": "Too many login attempts (rate limited)" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh JWT access token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "New access + refresh tokens" },
          "400": { "description": "Missing or invalid token type" },
          "401": { "description": "Expired or invalid refresh token" },
          "403": { "description": "Institution deactivated (staff only)" }
        }
      }
    },
...
    "/api/institutions": {
      "get": {
        "tags": ["Institutions"],
        "summary": "List all institutions (SUPER_ADMIN)",
        "responses": { "200": { "description": "List of institutions" } }
      },
      "post": {
        "tags": ["Institutions"],
        "summary": "Create institution (SUPER_ADMIN)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "location": { "type": "string" }, "level": { "type": "string", "enum": ["UNIVERSITET", "FAGSKOLE", "VGS", "UNGDOMSSKOLE"] } } } } } },
        "responses": { "201": { "description": "Institution created" }, "409": { "description": "Name already exists" } }
      }
    },
    "/api/institutions/{id}": {
      "put": {
        "tags": ["Institutions"],
        "summary": "Update institution (SUPER_ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "location": { "type": "string" }, "level": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["Institutions"],
        "summary": "Soft-delete institution (SUPER_ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
    "/api/institutions/inactive": {
      "get": {
        "tags": ["Institutions"],
        "summary": "List deactivated institutions (SUPER_ADMIN)",
        "responses": { "200": { "description": "List of deactivated institutions" }, "403": { "description": "Insufficient permissions" } }
      }
    },
    "/api/institutions/{id}/reactivate": {
      "put": {
        "tags": ["Institutions"],
        "summary": "Reactivate a deactivated institution (SUPER_ADMIN)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Institution reactivated" }, "403": { "description": "Insufficient permissions" }, "404": { "description": "Not found" } }
      }
    },
...
  },
...
}